% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Factor bind}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{a factor}

\item{b}{a factor}
}
\value{
a factor
}
\description{
Concatenate two factors, outputting a factor
}
\examples{
fbind(iris$Species[c(1, 51, 101)], PlantGrowth$group[c(1, 11, 21)])
}
